pkg: Add Basic Monitoring Support for Jobs
Jobs can now be added and updated through jobs/start and jobs/end
routes. The DB will be kept up to date - and also internal memory will
monitor those jobs.

Fallback, if the server application crashes is implemented such that it
fetches the unfinished jobs from the database.

Guard: Monitors Jobs, and checks timeout. If the Job ends or timeouts,
the Guard will send the event off to Bruce.

Bruce: Monitors finished events. If the event was the last one with this
commit hash of the open events, It will most likely post to gerrit (not
implemented yet).

Signed-off-by: Christian Walter <christian.walter@9elements.com>